chore: cargo clippy fixes and improve day12 part2

+++ src/year_2022/day_08.rs
@@ -22,7 +22,7 @@ impl Forest {
     }
 
     fn get(&self, row: usize, col: usize) -> Option<&Tree> {
-        self.trees.get(row as usize)?.get(col as usize)
+        self.trees.get(row)?.get(col)
     }
 }
